8361648: Update Glib to 2.84.3

+++ modules/javafx.media/src/main/native/gstreamer/3rd_party/glib/gobject/gvaluetypes.c
@@ -190,7 +190,8 @@ value_collect_float (GValue      *value,
          GTypeCValue *collect_values,
          guint        collect_flags)
 {
-  value->data[0].v_float = collect_values[0].v_double;
+  /* This necessarily loses precision */
+  value->data[0].v_float = (gfloat) collect_values[0].v_double;
 
   return NULL;
 }
